+++ PS7-x64/Modules/AdminToolbox.Office365/ChangeLog.md
@@ -26,3 +26,4 @@
 * **2.1.0.0** Added Functions Get-TenantDiagrams and Invoke-EXOSharedSession
 * **2.2.0.0** Added function Enable-MailboxAccess
 * **2.2.1.0** Updated module dependencies so that the Module AzViz doesn't install Az modules.
+* **2.2.2.0** Updates for output made by Enable-MailboxAccess